backend/requirements.txt
# Core Framework
fastapi==0.115.7
uvicorn[standard]==0.34.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.38
psycopg2-binary==2.9.11
alembic==1.13.1

# Geospatial
geoalchemy2==0.14.3
shapely==2.0.6
geopy==2.4.1

# Validation and Settings
pydantic==2.10.6
pydantic-settings==2.12.0
python-dotenv==1.0.0

# Security - HTML Sanitization
bleach==6.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing
python-multipart==0.0.6  # Form data
pyotp==2.9.0  # TOTP for 2FA
qrcode==7.4.2  # QR codes for 2FA setup

# Email Service (Brevo)
sib-api-v3-sdk==7.6.0  # Brevo (Sendinblue) API

# HTTP Client
httpx==0.26.0

# Background Tasks (Optional for MVP)
celery==5.3.4
redis==5.0.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
